Use the SDL2 software renderer
The way the SDL renderer is being used lends itself to horribly inefficient hardware rendering, even on an integrated GPU.

If we want a hardware renderer, I'd enable OpenGL and probably end up writing a shader.

This results in the blit step moving from the top of the profiling results to not being there at all.

EDIT: Never mind, software rendering *is* slower. Although it does seem to use 10 MB less RAM...